Split code into files - and all tests pass again
