+++ setup.py
@@ -3,7 +3,7 @@ import os
 from setuptools import setup
 
 
-with open(os.path.join('src', 'pydocstyle.py')) as f:
+with open(os.path.join('src', 'pydocstyle', 'pydocstyle.py')) as f:
     for line in f:
         if line.startswith('__version__'):
             version = eval(line.split('=')[-1])
@@ -27,12 +27,12 @@ setup(
         'License :: OSI Approved :: MIT License',
     ],
     keywords='pydocstyle, PEP 257, pep257, PEP 8, pep8, docstrings',
-    package_dir={'': 'src'},
-    py_modules=['pydocstyle'],
+    package_dir={'pydocstyle': 'src/pydocstyle'},
+    #py_modules=['pydocstyle'],
     entry_points={
         'console_scripts': [
-            'pydocstyle = pydocstyle:main',
-            'pep257 = pydocstyle:main_pep257',
+            'pydocstyle = pydocstyle.cli:main',
+            'pep257 = pydocstyle.cli:main_pep257',
         ],
     },
 )
