+++ src/pydocstyle/pydocstyle.py
@@ -1,0 +1,16 @@
+#! /usr/bin/env python
+"""Static analysis tool for checking docstring conventions and style.
+
+The repository is located at:
+http://github.com/PyCQA/pydocstyle
+
+"""
+from __future__ import with_statement
+
+import logging
+
+log = logging.getLogger(__name__)
+
+
+__version__ = '1.0.1-rc1'
+__all__ = ()
