+++ src/pydocstyle/tests/test_cases/multi_line_summary_start.py
