+++ src/pydocstyle/tests/test_cases/nested_class.py
