+++ src/pydocstyle/tests/test_decorators.py
@@ -2,6 +2,7 @@
 
 Use tox or py.test to run the test suite.
 """
+import checker
 
 try:
     from StringIO import StringIO
@@ -10,7 +11,7 @@ except ImportError:
 
 import textwrap
 
-from .. import pydocstyle
+from .. import pydocstyle, parser
 
 
 class TestParser(object):
@@ -23,7 +24,7 @@ class TestParser(object):
             class Foo:
                 pass
         """)
-        module = pydocstyle.parse(StringIO(code), 'dummy.py')
+        module = checker.parse(StringIO(code), 'dummy.py')
         decorators = module.children[0].decorators
 
         assert 1 == len(decorators)
@@ -43,7 +44,7 @@ class TestParser(object):
                 pass
         """)
 
-        module = pydocstyle.parse(StringIO(code), 'dummy.py')
+        module = checker.parse(StringIO(code), 'dummy.py')
         defined_class = module.children[0]
         decorators = defined_class.decorators
 
@@ -65,7 +66,7 @@ class TestParser(object):
                 class NestedClass:
                     pass
         """)
-        module = pydocstyle.parse(StringIO(code), 'dummy.py')
+        module = checker.parse(StringIO(code), 'dummy.py')
         nested_class = module.children[0].children[0]
         decorators = nested_class.decorators
 
@@ -82,7 +83,7 @@ class TestParser(object):
                     pass
         """)
 
-        module = pydocstyle.parse(StringIO(code), 'dummy.py')
+        module = checker.parse(StringIO(code), 'dummy.py')
         defined_class = module.children[0]
         decorators = defined_class.children[0].decorators
 
@@ -104,7 +105,7 @@ class TestParser(object):
                     pass
         """)
 
-        module = pydocstyle.parse(StringIO(code), 'dummy.py')
+        module = checker.parse(StringIO(code), 'dummy.py')
         defined_class = module.children[0]
         decorators = defined_class.children[0].decorators
 
@@ -124,7 +125,7 @@ class TestParser(object):
                 pass
         """)
 
-        module = pydocstyle.parse(StringIO(code), 'dummy.py')
+        module = checker.parse(StringIO(code), 'dummy.py')
         decorators = module.children[0].decorators
 
         assert 1 == len(decorators)
@@ -142,7 +143,7 @@ class TestParser(object):
                         pass
         """)
 
-        module = pydocstyle.parse(StringIO(code), 'dummy.py')
+        module = checker.parse(StringIO(code), 'dummy.py')
         defined_class = module.children[0]
         decorators = defined_class.children[0].children[0].decorators
 
@@ -163,19 +164,19 @@ class TestMethod:
                 def %s(self):
         """ % (name))
 
-        module = pydocstyle.Module('module_name', source, 0, 1, [],
-                                   'Docstring for module', [], None, all)
+        module = parser.Module('module_name', source, 0, 1, [],
+                               'Docstring for module', [], None, all)
 
-        cls = pydocstyle.Class('ClassName', source, 0, 1, [],
-                               'Docstring for class', children, module, all)
+        cls = parser.Class('ClassName', source, 0, 1, [],
+                           'Docstring for class', children, module, all)
 
-        return pydocstyle.Method(name, source, 0, 1, [],
-                                 'Docstring for method', children, cls, all)
+        return parser.Method(name, source, 0, 1, [],
+                             'Docstring for method', children, cls, all)
 
     def test_is_public_normal(self):
         """Methods are normally public, even if decorated."""
         method = self.makeMethod('methodName')
-        method.decorators = [pydocstyle.Decorator('some_decorator', [])]
+        method.decorators = [parser.Decorator('some_decorator', [])]
 
         assert method.is_public
 
@@ -183,8 +184,8 @@ class TestMethod:
         """Setter methods are considered private."""
         method = self.makeMethod('methodName')
         method.decorators = [
-            pydocstyle.Decorator('some_decorator', []),
-            pydocstyle.Decorator('methodName.setter', []),
+            parser.Decorator('some_decorator', []),
+            parser.Decorator('methodName.setter', []),
         ]
 
         assert not method.is_public
@@ -193,8 +194,8 @@ class TestMethod:
         """Deleter methods are also considered private."""
         method = self.makeMethod('methodName')
         method.decorators = [
-            pydocstyle.Decorator('methodName.deleter', []),
-            pydocstyle.Decorator('another_decorator', []),
+            parser.Decorator('methodName.deleter', []),
+            parser.Decorator('another_decorator', []),
         ]
 
         assert not method.is_public
@@ -203,8 +204,8 @@ class TestMethod:
         """Common prefix does not necessarily indicate private."""
         method = self.makeMethod("foo")
         method.decorators = [
-            pydocstyle.Decorator('foobar', []),
-            pydocstyle.Decorator('foobar.baz', []),
+            parser.Decorator('foobar', []),
+            parser.Decorator('foobar.baz', []),
         ]
 
         assert method.is_public
