+++ src/pydocstyle/utils.py
@@ -1,0 +1,2 @@
+def is_blank(string):
+    return not string.strip()
